Wallet now can be safely deleted only when it doesn't contain any operations.

+++ app/src/Controller/WalletController.php
@@ -173,6 +173,12 @@ class WalletController extends AbstractController
      */
     public function delete(Request $request, Wallet $wallet, WalletRepository $walletRepository): Response
     {
+        if ($wallet->getOperations()->count()) {
+            $this->addFlash('warning', 'message_category_contains_tasks');
+
+            return $this->redirectToRoute('wallet_index');
+        }
+
         $form = $this->createForm(FormType::class, $wallet, ['method' => 'DELETE']);
         $form->handleRequest($request);
